cribbage.cpp: Fix cut card selection, use uint64_t
`analyze_hand` was wrong regarding selecting the cut card.  We
were using C(46,3) to choose 3 cards: two cards for the
opponent's possible discard into the crib, and a third card for
the possible cut card.  In combinations, order doesn't matter,
but we were always using the third card for the cut card.  That
meant we were only considering C(46,3)=15'180 possibilities for
the crib but we should have been considering C(46,2)*44=45'540
possibilities (all of the combinations C(46,2) of choosing 2
cards from 46, and then the cut card can be any of the remaining
44 cards).

The other programming language implementations still have this
bug.

Represent a deck or hand of cards as a uint64_t bitfield instead
of an array.  Most operations on a deck/hand are in constant time
now, not linear with the number of cards in the deck/hand.  And
copying a Hand object is now cheaper (8 bytes instead of 52*4).